Thinned out helpers.py
